+++ dashboard/src/pages/library/anthology/show.tsx
@@ -1,7 +1,5 @@
 import { useParams } from "react-router-dom";
 import { Layout, Affix, Col, Row } from "antd";
-import HeadBar from "../../../components/library/HeadBar";
-import FooterBar from "../../../components/library/FooterBar";
 import AnthologyDetail from "../../../components/article/AnthologyDetail";
 
 const { Content, Header } = Layout;
@@ -15,37 +13,32 @@ const Widget = () => {
 	const pageMaxWidth = "1260px";
 	return (
 		<Layout>
-			<HeadBar selectedKeys="anthology" />
-			<Layout>
-				<Affix offsetTop={0}>
-					<Header style={{ backgroundColor: "gray", height: "3.5em" }}>
-						<Col flex="auto"></Col>
-						<Col flex={pageMaxWidth}>
-							<div>
-								{aid}@{channel}
-							</div>
-						</Col>
-						<Col flex="auto"></Col>
-					</Header>
-				</Affix>
+			<Affix offsetTop={0}>
+				<Header style={{ backgroundColor: "gray", height: "3.5em" }}>
+					<Col flex="auto"></Col>
+					<Col flex={pageMaxWidth}>
+						<div>
+							{aid}@{channel}
+						</div>
+					</Col>
+					<Col flex="auto"></Col>
+				</Header>
+			</Affix>
 
-				<Content>
-					<Row>
-						<Col flex="auto"></Col>
-						<Col flex={pageMaxWidth}>
-							<Row>
-								<Col span="18">
-									<AnthologyDetail aid={aid} />
-								</Col>
-								<Col span="6"></Col>
-							</Row>
-						</Col>
-						<Col flex="auto"></Col>
-					</Row>
-				</Content>
-			</Layout>
-
-			<FooterBar />
+			<Content>
+				<Row>
+					<Col flex="auto"></Col>
+					<Col flex={pageMaxWidth}>
+						<Row>
+							<Col span="18">
+								<AnthologyDetail aid={aid} />
+							</Col>
+							<Col span="6"></Col>
+						</Row>
+					</Col>
+					<Col flex="auto"></Col>
+				</Row>
+			</Content>
 		</Layout>
 	);
 };
